Fixed toggle so that the switch shows its checked when changing pages, also saves state to local storage so that the page remmebers the settings betweeen sessions and refreshes, couldn't get the toggle onto the loginpage as it is server rendered and thus doesn't recognise local storage
