+++ Components/hamburger.js
@@ -4,30 +4,11 @@ import { Button } from "@chakra-ui/react";
 import { Menu, MenuButton, MenuList, MenuItem, MenuDivider, Switch } from "@chakra-ui/react";
 import Link from "next/link";
 import styles from "../styles/Burger.module.css";
-import { useState } from "react";
+import ColourblindToggle from "./ColourblindToggle";
 
 
 
 const Hamburger = ({pet}) => {
-
-  const [colourBlind, setColourBlind] = useState(false)
-
-  function changeColour(e){
-    if (colourBlind === false){
-      var r = document.querySelector(':root');
-    var rs = getComputedStyle(r);
-    r.style.setProperty('--main-color', '#122746');
-    r.style.setProperty('--second-color', '#8ea3c6');
-    setColourBlind(true)
-    } else{
-      var r = document.querySelector(':root');
-    var rs = getComputedStyle(r);
-    r.style.setProperty('--main-color', '#448FFF');
-    r.style.setProperty('--second-color', '#B2D1FF');
-    setColourBlind(false)
-    }
-  }
-
 
   if(pet){
     return (
@@ -54,10 +35,7 @@ const Hamburger = ({pet}) => {
     <MenuItem>Useful Links</MenuItem>
   </Link>
   <MenuDivider />
-  <div className="colour-toggle">
-    <label>Colourblind Mode</label>
-    <Switch onChange={changeColour} value={colourBlind}></Switch>
-  </div>
+  <ColourblindToggle />
   <Link href= "/api/auth/logout">
     <MenuItem>Sign Out</MenuItem>
   </Link>
@@ -72,10 +50,7 @@ else{
     <Menu>
     <MenuButton as={Button} rightIcon={<HamburgerIcon />} className={styles.menubutton}/>
     <MenuList className={styles.menu}>
-    <div className="colour-toggle">
-    <label>Colourblind Mode</label>
-    <Switch onChange={changeColour} value={colourBlind}></Switch>
-  </div>
+    <ColourblindToggle />
   <Link href= "/api/auth/logout">
   <MenuItem>Sign Out</MenuItem>
 </Link>
